Change call host function return type from int32_t to int64_t

+++ libraries/chain/apply_context.cpp
@@ -247,7 +247,7 @@ void apply_context::exec()
 
 } /// exec()
 
-int32_t apply_context::execute_sync_call(name receiver, uint64_t flags, std::span<const char> data) {
+int64_t apply_context::execute_sync_call(name receiver, uint64_t flags, std::span<const char> data) {
    assert(sync_call_ctx.has_value() ^ (act != nullptr)); // can be only one of action and sync call
 
    dlog("receiver: ${r}, flags: ${f}, data size: ${s}",
